add section for making predictions in predict next purchase
