+++ docs/source/examples/demo/next_purchase/future.csv
@@ -1,0 +1,80 @@
+id,order_id,product_id,add_to_cart_order,reordered,product_name,aisle_id,department_id,department,user_id,order_time
+290784,2047,28155,1,1,Premium Lager Beer,27,5,alcohol,7622,2015-01-24 18:00:00
+290785,2047,22729,3,0,VO Canadian Whiskey,124,5,alcohol,7622,2015-01-24 18:00:00
+290786,13058,2120,2,1,Sauvignon Blanc,62,5,alcohol,54616,2015-01-31 15:00:00
+290787,13058,21948,1,1,Pinot Noir Wine,28,5,alcohol,54616,2015-01-31 15:00:00
+290788,26060,2120,3,1,Sauvignon Blanc,62,5,alcohol,26568,2015-01-09 15:00:00
+290789,26060,38444,2,1,Chardonnay,62,5,alcohol,26568,2015-01-09 15:00:00
+290790,26060,17859,1,1,Lager Beer,27,5,alcohol,26568,2015-01-09 15:00:00
+290791,77553,2120,2,0,Sauvignon Blanc,62,5,alcohol,12720,2015-01-01 12:00:00
+290792,77553,2022,4,0,Brut Cuvee Sparkling Champagne,134,5,alcohol,12720,2015-01-01 12:00:00
+290793,77553,40672,3,0,Brut California Champagne,134,5,alcohol,12720,2015-01-01 12:00:00
+290794,77553,17415,1,0,Chardonnay California 2009,62,5,alcohol,12720,2015-01-01 12:00:00
+290795,77553,43831,5,0,Extra Dry Champagne,134,5,alcohol,12720,2015-01-01 12:00:00
+290796,83883,2120,3,0,Sauvignon Blanc,62,5,alcohol,25379,2015-01-01 13:00:00
+290797,83883,36425,2,0,Chardonnay Wine,62,5,alcohol,25379,2015-01-01 13:00:00
+290798,83883,13162,1,0,Russian River Pinot Noir Wine,28,5,alcohol,25379,2015-01-01 13:00:00
+290799,96928,2120,1,1,Sauvignon Blanc,62,5,alcohol,10640,2015-01-24 20:00:00
+290800,96928,927,2,0,"Cabernet Sauvignon, Central Coast, 2011",28,5,alcohol,10640,2015-01-24 20:00:00
+290801,97176,2120,1,1,Sauvignon Blanc,62,5,alcohol,6941,2015-01-31 13:00:00
+290802,97176,31081,4,0,"Ale, India Pale, Brew Free! Or Die IPA",27,5,alcohol,6941,2015-01-31 13:00:00
+290803,97176,40087,3,0,12 Oz Beer,27,5,alcohol,6941,2015-01-31 13:00:00
+290804,97176,31113,2,1,Pinot Noir Rose,6,2,other,6941,2015-01-31 13:00:00
+290805,29321,36425,1,1,Chardonnay Wine,62,5,alcohol,3377,2015-01-03 17:00:00
+290806,29321,49544,2,1,Cabernet Sauvignon Wine,28,5,alcohol,3377,2015-01-03 17:00:00
+290807,29321,38251,3,1,Benchbreak Chardonnay,62,5,alcohol,3377,2015-01-03 17:00:00
+290808,37192,36425,4,0,Chardonnay Wine,62,5,alcohol,5504,2015-01-08 11:00:00
+290809,37192,28938,3,0,Refreshing White,62,5,alcohol,5504,2015-01-08 11:00:00
+290810,37192,44470,1,0,Backyard Variety Hard Strawberry Lemonade Hard Lemonade Hard Cranberry & Passion Fruit Lemonade Hard Tropical Mango Lemonade,27,5,alcohol,5504,2015-01-08 11:00:00
+290811,37192,16106,2,0,Budweiser,27,5,alcohol,5504,2015-01-08 11:00:00
+290812,11073,28578,2,1,Brown Ale,27,5,alcohol,38171,2015-01-11 11:00:00
+290813,11073,46088,4,1,Beer,27,5,alcohol,38171,2015-01-11 11:00:00
+290814,11073,17859,3,1,Lager Beer,27,5,alcohol,38171,2015-01-11 11:00:00
+290815,11073,31036,1,1,Crisp Hard Cider Crisp Apple,27,5,alcohol,38171,2015-01-11 11:00:00
+290816,24050,13303,1,1,"Pinot Noir, California",28,5,alcohol,53543,2015-01-06 14:00:00
+290817,24050,16753,2,1,Pinot Grigio Wine,62,5,alcohol,53543,2015-01-06 14:00:00
+290818,32604,13303,2,1,"Pinot Noir, California",28,5,alcohol,53543,2015-01-12 13:00:00
+290819,32604,12013,1,1,Pinot Noir,28,5,alcohol,53543,2015-01-12 13:00:00
+290820,27943,6878,1,0,Red Blend,28,5,alcohol,7819,2015-01-10 20:00:00
+290821,27943,10702,2,0,California Red Wine,28,5,alcohol,7819,2015-01-10 20:00:00
+290822,27943,38569,4,0,Old Vine Zinfandel Wine,28,5,alcohol,7819,2015-01-10 20:00:00
+290823,27943,837,3,0,California Merlot,28,5,alcohol,7819,2015-01-10 20:00:00
+290825,40972,46088,2,1,Beer,27,5,alcohol,42899,2015-01-06 14:00:00
+290826,40972,16139,1,1,Clara,27,5,alcohol,42899,2015-01-06 14:00:00
+290827,40972,20775,3,1,Grapefruit Sculpin IPA,27,5,alcohol,42899,2015-01-06 14:00:00
+290828,63592,38444,1,1,Chardonnay,62,5,alcohol,29098,2015-01-10 10:00:00
+290829,63592,3497,2,1,"Grape Wine, Delicious Red",28,5,alcohol,29098,2015-01-10 10:00:00
+290831,83070,38444,3,0,Chardonnay,62,5,alcohol,4120,2015-01-29 14:00:00
+290832,83070,12013,4,1,Pinot Noir,28,5,alcohol,4120,2015-01-29 14:00:00
+290833,83070,14385,2,0,Amber Ale,27,5,alcohol,4120,2015-01-29 14:00:00
+290834,83070,41539,1,0,Variety Pack Hard Cider,27,5,alcohol,4120,2015-01-29 14:00:00
+290835,65887,33153,2,1,IPA,27,5,alcohol,18775,2015-01-12 17:00:00
+290836,65887,7020,1,1,90 Minute Imperial Ipa,27,5,alcohol,18775,2015-01-12 17:00:00
+290838,34608,12013,2,0,Pinot Noir,28,5,alcohol,21275,2015-01-09 10:00:00
+290839,34608,1097,3,0,Sonoma County Sauvignon Blanc,62,5,alcohol,21275,2015-01-09 10:00:00
+290840,34608,31081,4,1,"Ale, India Pale, Brew Free! Or Die IPA",27,5,alcohol,21275,2015-01-09 10:00:00
+290841,34608,48710,1,0,"Pinto Noir, Monterey, 2007",28,5,alcohol,21275,2015-01-09 10:00:00
+290842,41240,12013,3,1,Pinot Noir,28,5,alcohol,24644,2015-02-01 16:00:00
+290843,41240,22690,1,0,Draught Beer,27,5,alcohol,24644,2015-02-01 16:00:00
+290844,41240,21974,2,1,Old Vines Zinfandel Wine,28,5,alcohol,24644,2015-02-01 16:00:00
+290845,99339,12013,1,1,Pinot Noir,28,5,alcohol,44816,2015-01-04 15:00:00
+290846,99339,13707,4,1,Essential Red,28,5,alcohol,44816,2015-01-04 15:00:00
+290847,99339,48939,3,0,Extra IPA Beer,27,5,alcohol,44816,2015-01-04 15:00:00
+290848,99339,18527,2,1,Indian Pale Ale,27,5,alcohol,44816,2015-01-04 15:00:00
+290849,104946,12013,1,0,Pinot Noir,28,5,alcohol,21204,2015-01-10 15:00:00
+290850,104946,28966,2,0,Blanc De Noir Champagne,134,5,alcohol,21204,2015-01-10 15:00:00
+290860,50531,48939,2,0,Extra IPA Beer,27,5,alcohol,59789,2015-01-13 15:00:00
+290861,50531,28238,1,1,Original Belgian Wheat Beer,27,5,alcohol,59789,2015-01-13 15:00:00
+290862,10476,40364,1,0,Draught,27,5,alcohol,53020,2015-01-04 11:00:00
+290863,10476,25456,2,0,Extra Stout Beer,27,5,alcohol,53020,2015-01-04 11:00:00
+290864,59632,26056,1,0,Old Rasputin Stout Beer,27,5,alcohol,21222,2015-01-09 16:00:00
+290865,59632,27840,2,0,Icelandic White Ale 6 Pack,27,5,alcohol,21222,2015-01-09 16:00:00
+290866,91111,15248,3,1,India Pale Ale Racer 5,27,5,alcohol,42191,2015-01-01 11:00:00
+290867,91111,32838,2,1,Belgian Style Tripel Ale,27,5,alcohol,42191,2015-01-01 11:00:00
+290868,91111,12527,1,0,Sidecar Orange Pale Ale,100,21,missing,42191,2015-01-01 11:00:00
+290870,97144,4079,1,0,Tequila Anejo,124,5,alcohol,54353,2015-01-11 10:00:00
+290871,97144,20135,2,0,"Cognac, Privilege V.S.O.P",124,5,alcohol,54353,2015-01-11 10:00:00
+290872,31925,22199,1,0,Extra-Dry Cava,134,5,alcohol,47329,2015-01-13 10:00:00
+290873,31925,24968,2,1,Sonoma Brut Sparkling Wine,134,5,alcohol,47329,2015-01-13 10:00:00
+290877,62363,1000,2,1,Apricots,18,10,bulk,59269,2015-01-06 06:00:00
+290878,62363,5161,1,1,Dried Mango,18,10,bulk,59269,2015-01-06 06:00:00
